backend/requirements.txt
dj-rest-auth==2.2.3
Django==4.0.3
django-allauth==0.49.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
sqlparse==0.4.2
